+++ src/cpp/core/system/file_monitor/FileMonitor.cpp
@@ -420,7 +420,8 @@ public:
 
 public:
    RegistrationCommand()
-      : type_(None)
+      : type_(None),
+        recursive_(false)
    {
    }